# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bakery website API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Bakery product operations
  - name: orders
    description: Order operations
  - name: reservations
    description: Table reservation operations
  - name: reviews
    description: Customer review operations
  - name: contact
    description: Contact form operations
  - name: auth
    description: Authentication operations
  - name: admin
    description: Admin dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # Auth
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Products
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /products/categories:
    get:
      operationId: listCategories
      tags: [products]
      summary: List product categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin or own)
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Reservations
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      responses:
        "200":
          description: Reservations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Book a table reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get reservation by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    patch:
      operationId: updateReservationStatus
      tags: [reservations]
      summary: Update reservation status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationStatusUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  # Reviews
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all published reviews
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/{id}/approve:
    patch:
      operationId: approveReview
      tags: [reviews]
      summary: Approve review (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Contact
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /contact/messages:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List contact messages (admin)
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"

  # Admin dashboard
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [admin]
      summary: Get recent orders for dashboard
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/sales-by-category:
    get:
      operationId: getSalesByCategory
      tags: [admin]
      summary: Get sales grouped by product category
      responses:
        "200":
          description: Sales by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySales"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, admin]
        createdAt:
          type: string
          format: date-time

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # Products
    Product:
      type: object
      required: [id, name, description, price, category, imageUrl, featured, available, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        available:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    ProductInput:
      type: object
      required: [name, description, price, category, imageUrl]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        available:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        available:
          type: boolean

    # Orders
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number

    Order:
      type: object
      required: [id, customerName, customerEmail, customerPhone, deliveryAddress, items, subtotal, total, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        createdAt:
          type: string
          format: date-time

    OrderInput:
      type: object
      required: [customerName, customerEmail, customerPhone, deliveryAddress, items]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
        couponCode:
          type: string
        notes:
          type: string

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]

    # Reservations
    Reservation:
      type: object
      required: [id, customerName, customerEmail, customerPhone, date, time, guestCount, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
        specialRequests:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string
          format: date-time

    ReservationInput:
      type: object
      required: [customerName, customerEmail, customerPhone, date, time, guestCount]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
          minimum: 1
          maximum: 20
        specialRequests:
          type: string

    ReservationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]

    # Reviews
    Review:
      type: object
      required: [id, customerName, rating, comment, approved, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time

    ReviewInput:
      type: object
      required: [customerName, rating, comment]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # Contact
    ContactMessage:
      type: object
      required: [id, name, email, subject, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        phone:
          type: ["string", "null"]
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        phone:
          type: string

    # Admin
    AdminStats:
      type: object
      required: [totalOrders, totalRevenue, totalReservations, totalCustomers, pendingOrders, todayOrders, todayRevenue, averageRating]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        totalReservations:
          type: integer
        totalCustomers:
          type: integer
        pendingOrders:
          type: integer
        todayOrders:
          type: integer
        todayRevenue:
          type: number
        averageRating:
          type: number

    CategorySales:
      type: object
      required: [category, totalSold, revenue]
      properties:
        category:
          type: string
        totalSold:
          type: integer
        revenue:
          type: number