# lib/api-zod/src/generated/types/categorySales.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery website API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySales {
  category: string;
  totalSold: number;
  revenue: number;
}