# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery website API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  items: OrderItem[];
  subtotal: number;
  discount?: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  notes?: string | null;
  status: OrderStatus;
  createdAt: Date;
}